server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
httpx==0.27.0
python-json-logger==2.0.7
APScheduler==3.10.4
SQLAlchemy==2.0.30
alembic==1.13.1
PyYAML==6.0.2

